refactor(wash-runtime): split the nats plugin out of its module root
`wasmcloud_nats/mod.rs` carried three jobs at once: the module root and its
bindgen worlds, the parsing of what a binding is described by, and the plugin
itself. Every sibling in the directory is one concern, and the plugin dirs
that fan out into submodules -- `wasi_keyvalue`, `wasi_blobstore` -- keep the
`HostPlugin` impl in a file of its own and the root thin.

So: `WasmcloudNats`, `ComponentData` and the impls move to `plugin.rs`, the
subscription structs and their parsers join the config they parse in
`config.rs`, and each test follows its subject. The root is left with the
module declarations, the three bindgen worlds and the two consts.

Move only -- no behaviour change. Everything outside the module keeps its
path: `WasmcloudNats` and `ComponentData` are re-exported from the root.
Beyond the relocation the diff is nine lines: two imports the root no longer
needs, four parsers widened to `pub(super)` now that they are reached from a
sibling, and three doc links respelled from the module they moved out of.
